feat: add validation

+++ backend/src/products/products.routes.ts
@@ -1,29 +1,47 @@
 import { Request, Response, Router } from 'express'
 import { deleteProduct, getProduct, insertProduct, listProducts, updateProduct } from './products.repository';
+import { createProductSchema, updateProductSchema } from './products.schema';
+import { validationSchemaMiddleware } from '../middleware';
 
 export const productsRouter = Router();
 
-productsRouter.post('/products', async (req: Request, res: Response) => {
-  const product = req.body;
-
-  await insertProduct(product);
-
-  res.sendStatus(201);
+productsRouter.post(
+  '/products',
+  validationSchemaMiddleware(createProductSchema),
+  async (req: Request, res: Response) => {
+    const product = req.body;
+    await insertProduct(product);
+    res.sendStatus(201);
 });
 
-productsRouter.put('/products/:id', async (req: Request, res: Response) => {
+productsRouter.put(
+  '/products/:id',
+  validationSchemaMiddleware(updateProductSchema),
+  async (req: Request, res: Response) => {
 
-  const productId = Number(req.params.id);
-  const productData = req.body;
+    const productId = Number(req.params.id);
+    const productData = req.body;
 
-  await updateProduct(productId, productData)
+    const product = await getProduct(productId);
 
-  res.sendStatus(204);
+    if (!product) {
+      return res.sendStatus(404);
+    }
+
+    await updateProduct(productId, productData)
+    res.sendStatus(204);
+
 });
 
 
 productsRouter.delete('/products/:id', async (req: Request, res: Response) => {
   const productId = Number(req.params.id);
+
+  const product = await getProduct(productId);
+
+  if (!product) {
+    return res.sendStatus(404);
+  }
 
   await deleteProduct(productId)
 
@@ -33,6 +51,10 @@ productsRouter.delete('/products/:id', async (req: Request, res: Response) => {
 productsRouter.get('/products/:id', async (req: Request, res: Response) => {
   const productId = Number(req.params.id);
   const product = await getProduct(productId)
+
+  if (!product) {
+    return res.sendStatus(404);
+  }
 
   res.status(200).json(product);
 });
